feat: changing folder structure
