+++ src/main/evaluate-expression/index.ts
@@ -1,0 +1,4 @@
+import evaluateExpression from "./evaluate-expression";
+
+export { operate } from "./evaluate-expression";
+export default evaluateExpression;
